backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
Pillow==10.1.0
python-dateutil==2.8.2
geopy==2.4.1
gunicorn==21.2.0
